+++ lib/live_activities_platform_interface.dart
@@ -1,9 +1,9 @@
+import 'package:live_activities/live_activities_method_channel.dart';
 import 'package:live_activities/models/activity_update.dart';
+import 'package:live_activities/models/alert_config.dart';
 import 'package:live_activities/models/live_activity_state.dart';
 import 'package:live_activities/models/url_scheme_data.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
-
-import 'live_activities_method_channel.dart';
 
 abstract class LiveActivitiesPlatform extends PlatformInterface {
   /// Constructs a LiveActivitiesPlatform.
@@ -26,7 +26,10 @@ abstract class LiveActivitiesPlatform extends PlatformInterface {
     _instance = instance;
   }
 
-  Future init(String appGroupId) {
+  Future init(
+    String appGroupId, {
+    String? urlScheme,
+  }) {
     throw UnimplementedError('init() has not been implemented.');
   }
 
@@ -38,7 +41,8 @@ abstract class LiveActivitiesPlatform extends PlatformInterface {
     throw UnimplementedError('createActivity() has not been implemented.');
   }
 
-  Future updateActivity(String activityId, Map<String, dynamic> data) {
+  Future updateActivity(String activityId, Map<String, dynamic> data,
+      [AlertConfig? alertConfig]) {
     throw UnimplementedError('updateActivity() has not been implemented.');
   }
 
@@ -54,6 +58,10 @@ abstract class LiveActivitiesPlatform extends PlatformInterface {
     throw UnimplementedError('endAllActivities() has not been implemented.');
   }
 
+  Future<Map<String, LiveActivityState>> getAllActivities() {
+    throw UnimplementedError('getAllActivities() has not been implemented.');
+  }
+
   Future<bool> areActivitiesEnabled() {
     throw UnimplementedError(
         'areActivitiesEnabled() has not been implemented.');
@@ -63,7 +71,7 @@ abstract class LiveActivitiesPlatform extends PlatformInterface {
     throw UnimplementedError('urlSchemeStream() has not been implemented.');
   }
 
-  Future<LiveActivityState> getActivityState(String activityId) {
+  Future<LiveActivityState?> getActivityState(String activityId) {
     throw UnimplementedError('getActivityState() has not been implemented.');
   }