+++ CHANGELOG.md
@@ -1,3 +1,46 @@
+## 1.9.1
+
+* ✨ Add update with alert config (thanks @charlesRmajor 👍).
+* ✨ Add an option to use preloaded images (thanks @Niklas-Sommer 👍).
+* ✨ Add Android support - currently only used to check if live activities is supported (thanks @ggirotto 👍).
+* ✨ Example app support Material 3.
+* 🐛 Fix tests.
+* 📝 Update README.md.
+* ⬆️ Upgrade dependencies.
+
+## 1.9.0
+
+- ✨ **BREAKING CHANGE**: Add the ability to handle multiple live notification (thanks @Clon1998 👍).
+
+Please follow this tutorial to add implement it:
+
+- Add the following Swift extension at the end of your extension code:
+
+```swift
+extension LiveActivitiesAppAttributes {
+  func prefixedKey(_ key: String) -> String {
+    return "\(id)_\(key)"
+  }
+}
+```
+
+- For each keys on your native Swift code, please changes the following lines:
+
+```swift
+let myVariableFromFlutter = sharedDefault.string(forKey: "myVariableFromFlutter") // repleace this by ...
+let myVariableFromFlutter = sharedDefault.string(forKey: context.attributes.prefixedKey("myVariableFromFlutter")) // <-- this
+```
+ 
+- 🐛 Fix stall state for unknown activityId (thanks @Clon1998 👍).
+- 🐛 Now return `null` value when activity is not found in `getActivityState()`.
+
+## 1.8.0
+
+* ✨ Add url scheme optional argument.
+* ✨ Add sinks unregister on engine end (thanks @ggirotto 👍).
+* 🐛 Fix example images size.
+* ⬆️ Upgrade dependencies.
+
 ## 1.7.5
 
 * 🚨 Lint some code.